project stop

+++ src/components/Modal/DelModal.jsx
@@ -2,12 +2,17 @@ import React, { useState } from 'react';
 import './DelModal.css'
 import { useInfoContext } from '../../context/Context';
 
-const DeleteModal = ({ onDelete }) => {
-    const {setShowModal, showModal, exit} = useInfoContext()
+const DeleteModal = ({ onDelete, chatDelete }) => {
+    const {setShowModal, showModal} = useInfoContext()
 
   const handleDelete = () => {
-    onDelete(); // Call the delete function passed from the parent component
-    setShowModal(false); // Close the modal after deletion
+    if(onDelete){
+      onDelete();
+    };
+    if(chatDelete){
+      chatDelete();
+    };
+    setShowModal(false);
   };
 
   return (
@@ -17,8 +22,10 @@ const DeleteModal = ({ onDelete }) => {
           <div className="modal-content delete-modal">
             <h2>Confirm Deletion</h2>
             <p>Are you sure you want to delete?</p>
-            <button style={{color: 'red'}} onClick={handleDelete}>Delete</button>
-            <button onClick={() => setShowModal(false)}>Cancel</button>
+            <div className='del-button'>
+              <button style={{color: 'red'}} onClick={handleDelete}>Delete</button>
+              <button onClick={() => setShowModal(false)}>Cancel</button>
+            </div>
           </div>
         </div>
       )}
